Allow setting query trace config options
The following allows the updating of the query trace options on the
controller config. As the database is controller only we only need to
set it there.

The controller config is a bit finicky as we have to ensure it's being
correctly set at every level. When changing the value it will cause the
agent to bounce, which will force it to pick the updated value up.

Tests will be required to ensure we have good coverage, along with a bug
around setting times in the config directly.
