+++ agent/agent.go
@@ -12,6 +12,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"strconv"
+	"time"
 
 	"github.com/juju/collections/set"
 	"github.com/juju/errors"
@@ -286,6 +287,14 @@ type Config interface {
 	// AgentLogfileMaxBackups returns the number of old agent/controller log
 	// files to keep (compressed).
 	AgentLogfileMaxBackups() int
+
+	// QueryTracingEnabled returns whether query tracing is enabled.
+	QueryTracingEnabled() bool
+
+	// QueryTracingThreshold returns the threshold for query tracing. The
+	// lower the threshold, the more queries will be output. A value of 0
+	// means all queries will be output.
+	QueryTracingThreshold() time.Duration
 }
 
 type configSetterOnly interface {
@@ -333,6 +342,12 @@ type configSetterOnly interface {
 
 	// SetLoggingConfig sets the logging config value for the agent.
 	SetLoggingConfig(string)
+
+	// SetQueryTracingEnabled sets whether query tracing is enabled.
+	SetQueryTracingEnabled(bool)
+
+	// SetQueryTracingThreshold sets the threshold for query tracing.
+	SetQueryTracingThreshold(time.Duration)
 }
 
 // LogFileName returns the filename for the Agent's log file.
@@ -408,6 +423,8 @@ type configInternal struct {
 	jujuDBSnapChannel      string
 	agentLogfileMaxSizeMB  int
 	agentLogfileMaxBackups int
+	queryTracingEnabled    bool
+	queryTracingThreshold  time.Duration
 }
 
 // AgentConfigParams holds the parameters required to create
@@ -428,6 +445,8 @@ type AgentConfigParams struct {
 	JujuDBSnapChannel      string
 	AgentLogfileMaxSizeMB  int
 	AgentLogfileMaxBackups int
+	QueryTracingEnabled    bool
+	QueryTracingThreshold  time.Duration
 }
 
 // NewAgentConfig returns a new config object suitable for use for a
@@ -491,6 +510,8 @@ func NewAgentConfig(configParams AgentConfigParams) (ConfigSetterWriter, error) 
 		jujuDBSnapChannel:      configParams.JujuDBSnapChannel,
 		agentLogfileMaxSizeMB:  configParams.AgentLogfileMaxSizeMB,
 		agentLogfileMaxBackups: configParams.AgentLogfileMaxBackups,
+		queryTracingEnabled:    configParams.QueryTracingEnabled,
+		queryTracingThreshold:  configParams.QueryTracingThreshold,
 	}
 	if len(configParams.APIAddresses) > 0 {
 		config.apiDetails = &apiDetails{
@@ -811,6 +832,26 @@ func (c *configInternal) AgentLogfileMaxSizeMB() int {
 // AgentLogfileMaxBackups implements Config.
 func (c *configInternal) AgentLogfileMaxBackups() int {
 	return c.agentLogfileMaxBackups
+}
+
+// QueryTracingEnabled implements Config.
+func (c *configInternal) QueryTracingEnabled() bool {
+	return c.queryTracingEnabled
+}
+
+// SetQueryTracingEnabled implements configSetterOnly.
+func (c *configInternal) SetQueryTracingEnabled(v bool) {
+	c.queryTracingEnabled = v
+}
+
+// QueryTracingThreshold implements Config.
+func (c *configInternal) QueryTracingThreshold() time.Duration {
+	return c.queryTracingThreshold
+}
+
+// SetQueryTracingThreshold implements configSetterOnly.
+func (c *configInternal) SetQueryTracingThreshold(v time.Duration) {
+	c.queryTracingThreshold = v
 }
 
 var validAddr = regexp.MustCompile("^.+:[0-9]+$")
