+++ controller/config_test.go
@@ -184,221 +184,238 @@ var newConfigTests = []struct {
 		controller.CAASImageRepo: "foo?bar",
 	},
 	expectError: `docker image path "foo\?bar": invalid reference format`,
-},
-	{
-		about: "empty CAAS docker image repo",
-		config: controller.Config{
-			controller.CAASImageRepo: `{"foo": "bar"}`,
-		},
-		expectError: `empty repository not valid`,
-	}, {
-		about: "invalid CAAS operator docker image repo - leading colon",
-		config: controller.Config{
-			controller.CAASImageRepo: ":foo",
-		},
-		expectError: `docker image path ":foo": invalid reference format`,
-	}, {
-		about: "invalid CAAS docker image repo - trailing colon",
-		config: controller.Config{
-			controller.CAASImageRepo: `{"foo":""}`,
-		},
-		expectError: `empty repository not valid`,
-	}, {
-		about: "invalid CAAS docker image repo - extra colon",
-		config: controller.Config{
-			controller.CAASImageRepo: "foo::bar",
-		},
-		expectError: `docker image path "foo::bar": invalid reference format`,
-	}, {
-		about: "invalid CAAS docker image repo - leading /",
-		config: controller.Config{
-			controller.CAASImageRepo: "/foo",
-		},
-		expectError: `docker image path "/foo": invalid reference format`,
-	}, {
-		about: "invalid CAAS docker image repo - extra /",
-		config: controller.Config{
-			controller.CAASImageRepo: "foo//bar",
-		},
-		expectError: `docker image path "foo//bar": invalid reference format`,
-	}, {
-		about: "negative controller-api-port",
-		config: controller.Config{
-			controller.ControllerAPIPort: -5,
-		},
-		expectError: `non-positive integer for controller-api-port not valid`,
-	}, {
-		about: "controller-api-port matching api-port",
-		config: controller.Config{
-			controller.APIPort:           12345,
-			controller.ControllerAPIPort: 12345,
-		},
-		expectError: `controller-api-port matching api-port not valid`,
-	}, {
-		about: "controller-api-port matching state-port",
-		config: controller.Config{
-			controller.APIPort:           12345,
-			controller.StatePort:         54321,
-			controller.ControllerAPIPort: 54321,
-		},
-		expectError: `controller-api-port matching state-port not valid`,
-	}, {
-		about: "api-port-open-delay not a duration",
-		config: controller.Config{
-			controller.APIPortOpenDelay: "15",
-		},
-		expectError: `api-port-open-delay: conversion to duration: time: missing unit in duration "15"`,
-	}, {
-		about: "txn-prune-sleep-time not a duration",
-		config: controller.Config{
-			controller.PruneTxnSleepTime: "15",
-		},
-		expectError: `prune-txn-sleep-time: conversion to duration: time: missing unit in duration "15"`,
-	}, {
-		about: "mongo-memory-profile not valid",
-		config: controller.Config{
-			controller.MongoMemoryProfile: "not-valid",
-		},
-		expectError: `mongo-memory-profile: expected one of "low" or "default" got string\("not-valid"\)`,
-	}, {
-		about: "max-debug-log-duration not valid",
-		config: controller.Config{
-			controller.MaxDebugLogDuration: time.Duration(0),
-		},
-		expectError: `max-debug-log-duration cannot be zero`,
-	}, {
-		about: "agent-logfile-max-backups not valid",
-		config: controller.Config{
-			controller.AgentLogfileMaxBackups: -1,
-		},
-		expectError: `negative agent-logfile-max-backups not valid`,
-	}, {
-		about: "agent-logfile-max-size not valid",
-		config: controller.Config{
-			controller.AgentLogfileMaxSize: "0",
-		},
-		expectError: `agent-logfile-max-size less than 1 MB not valid`,
-	}, {
-		about: "model-logfile-max-backups not valid",
-		config: controller.Config{
-			controller.ModelLogfileMaxBackups: -1,
-		},
-		expectError: `negative model-logfile-max-backups not valid`,
-	}, {
-		about: "model-logfile-max-size not valid",
-		config: controller.Config{
-			controller.ModelLogfileMaxSize: "0",
-		},
-		expectError: `model-logfile-max-size less than 1 MB not valid`,
-	}, {
-		about: "agent-ratelimit-max non-int",
-		config: controller.Config{
-			controller.AgentRateLimitMax: "ten",
-		},
-		expectError: `agent-ratelimit-max: expected number, got string\("ten"\)`,
-	}, {
-		about: "agent-ratelimit-max negative",
-		config: controller.Config{
-			controller.AgentRateLimitMax: "-5",
-		},
-		expectError: `negative agent-ratelimit-max \(-5\) not valid`,
-	}, {
-		about: "agent-ratelimit-rate missing unit",
-		config: controller.Config{
-			controller.AgentRateLimitRate: "150",
-		},
-		expectError: `agent-ratelimit-rate: conversion to duration: time: missing unit in duration "?150"?`,
-	}, {
-		about: "agent-ratelimit-rate bad type, int",
-		config: controller.Config{
-			controller.AgentRateLimitRate: 150,
-		},
-		expectError: `agent-ratelimit-rate: expected string or time.Duration, got int\(150\)`,
-	}, {
-		about: "agent-ratelimit-rate zero",
-		config: controller.Config{
-			controller.AgentRateLimitRate: "0s",
-		},
-		expectError: `agent-ratelimit-rate cannot be zero`,
-	}, {
-		about: "agent-ratelimit-rate negative",
-		config: controller.Config{
-			controller.AgentRateLimitRate: "-5s",
-		},
-		expectError: `agent-ratelimit-rate cannot be negative`,
-	}, {
-		about: "agent-ratelimit-rate too large",
-		config: controller.Config{
-			controller.AgentRateLimitRate: "4h",
-		},
-		expectError: `agent-ratelimit-rate must be between 0..1m`,
-	}, {
-		about: "max-charm-state-size non-int",
-		config: controller.Config{
-			controller.MaxCharmStateSize: "ten",
-		},
-		expectError: `max-charm-state-size: expected number, got string\("ten"\)`,
-	}, {
-		about: "max-charm-state-size cannot be negative",
-		config: controller.Config{
-			controller.MaxCharmStateSize: "-42",
-		},
-		expectError: `invalid max charm state size: should be a number of bytes \(or 0 to disable limit\), got -42`,
-	}, {
-		about: "max-agent-state-size non-int",
-		config: controller.Config{
-			controller.MaxAgentStateSize: "ten",
-		},
-		expectError: `max-agent-state-size: expected number, got string\("ten"\)`,
-	}, {
-		about: "max-agent-state-size cannot be negative",
-		config: controller.Config{
-			controller.MaxAgentStateSize: "-42",
-		},
-		expectError: `invalid max agent state size: should be a number of bytes \(or 0 to disable limit\), got -42`,
-	}, {
-		about: "combined charm/agent state cannot exceed mongo's 16M limit/doc",
-		config: controller.Config{
-			controller.MaxCharmStateSize: "14000000",
-			controller.MaxAgentStateSize: "3000000",
-		},
-		expectError: `invalid max charm/agent state sizes: combined value should not exceed mongo's 16M per-document limit, got 17000000`,
-	}, {
-		about: "public-dns-address: expect string, got number",
-		config: controller.Config{
-			controller.PublicDNSAddress: 42,
-		},
-		expectError: `public-dns-address: expected string, got int\(42\)`,
-	}, {
-		about: "migration-agent-wait-time not a duration",
-		config: controller.Config{
-			controller.MigrationMinionWaitMax: "15",
-		},
-		expectError: `migration-agent-wait-time: conversion to duration: time: missing unit in duration "15"`,
-	}, {
-		about: "application-resource-download-limit cannot be negative",
-		config: controller.Config{
-			controller.ApplicationResourceDownloadLimit: "-42",
-		},
-		expectError: `negative application-resource-download-limit \(-42\) not valid, use 0 to disable the limit`,
-	}, {
-		about: "controller-resource-download-limit cannot be negative",
-		config: controller.Config{
-			controller.ControllerResourceDownloadLimit: "-42",
-		},
-		expectError: `negative controller-resource-download-limit \(-42\) not valid, use 0 to disable the limit`,
-	}, {
-		about: "login token refresh url",
-		config: controller.Config{
-			controller.LoginTokenRefreshURL: `https://xxxx`,
-		},
-	}, {
-		about: "invalid login token refresh url",
-		config: controller.Config{
-			controller.LoginTokenRefreshURL: `xxxx`,
-		},
-		expectError: `logic token refresh URL "xxxx" not valid`,
-	}, {}}
+}, {
+	about: "empty CAAS docker image repo",
+	config: controller.Config{
+		controller.CAASImageRepo: `{"foo": "bar"}`,
+	},
+	expectError: `empty repository not valid`,
+}, {
+	about: "invalid CAAS operator docker image repo - leading colon",
+	config: controller.Config{
+		controller.CAASImageRepo: ":foo",
+	},
+	expectError: `docker image path ":foo": invalid reference format`,
+}, {
+	about: "invalid CAAS docker image repo - trailing colon",
+	config: controller.Config{
+		controller.CAASImageRepo: `{"foo":""}`,
+	},
+	expectError: `empty repository not valid`,
+}, {
+	about: "invalid CAAS docker image repo - extra colon",
+	config: controller.Config{
+		controller.CAASImageRepo: "foo::bar",
+	},
+	expectError: `docker image path "foo::bar": invalid reference format`,
+}, {
+	about: "invalid CAAS docker image repo - leading /",
+	config: controller.Config{
+		controller.CAASImageRepo: "/foo",
+	},
+	expectError: `docker image path "/foo": invalid reference format`,
+}, {
+	about: "invalid CAAS docker image repo - extra /",
+	config: controller.Config{
+		controller.CAASImageRepo: "foo//bar",
+	},
+	expectError: `docker image path "foo//bar": invalid reference format`,
+}, {
+	about: "negative controller-api-port",
+	config: controller.Config{
+		controller.ControllerAPIPort: -5,
+	},
+	expectError: `non-positive integer for controller-api-port not valid`,
+}, {
+	about: "controller-api-port matching api-port",
+	config: controller.Config{
+		controller.APIPort:           12345,
+		controller.ControllerAPIPort: 12345,
+	},
+	expectError: `controller-api-port matching api-port not valid`,
+}, {
+	about: "controller-api-port matching state-port",
+	config: controller.Config{
+		controller.APIPort:           12345,
+		controller.StatePort:         54321,
+		controller.ControllerAPIPort: 54321,
+	},
+	expectError: `controller-api-port matching state-port not valid`,
+}, {
+	about: "api-port-open-delay not a duration",
+	config: controller.Config{
+		controller.APIPortOpenDelay: "15",
+	},
+	expectError: `api-port-open-delay: conversion to duration: time: missing unit in duration "15"`,
+}, {
+	about: "txn-prune-sleep-time not a duration",
+	config: controller.Config{
+		controller.PruneTxnSleepTime: "15",
+	},
+	expectError: `prune-txn-sleep-time: conversion to duration: time: missing unit in duration "15"`,
+}, {
+	about: "mongo-memory-profile not valid",
+	config: controller.Config{
+		controller.MongoMemoryProfile: "not-valid",
+	},
+	expectError: `mongo-memory-profile: expected one of "low" or "default" got string\("not-valid"\)`,
+}, {
+	about: "max-debug-log-duration not valid",
+	config: controller.Config{
+		controller.MaxDebugLogDuration: time.Duration(0),
+	},
+	expectError: `max-debug-log-duration cannot be zero`,
+}, {
+	about: "agent-logfile-max-backups not valid",
+	config: controller.Config{
+		controller.AgentLogfileMaxBackups: -1,
+	},
+	expectError: `negative agent-logfile-max-backups not valid`,
+}, {
+	about: "agent-logfile-max-size not valid",
+	config: controller.Config{
+		controller.AgentLogfileMaxSize: "0",
+	},
+	expectError: `agent-logfile-max-size less than 1 MB not valid`,
+}, {
+	about: "model-logfile-max-backups not valid",
+	config: controller.Config{
+		controller.ModelLogfileMaxBackups: -1,
+	},
+	expectError: `negative model-logfile-max-backups not valid`,
+}, {
+	about: "model-logfile-max-size not valid",
+	config: controller.Config{
+		controller.ModelLogfileMaxSize: "0",
+	},
+	expectError: `model-logfile-max-size less than 1 MB not valid`,
+}, {
+	about: "agent-ratelimit-max non-int",
+	config: controller.Config{
+		controller.AgentRateLimitMax: "ten",
+	},
+	expectError: `agent-ratelimit-max: expected number, got string\("ten"\)`,
+}, {
+	about: "agent-ratelimit-max negative",
+	config: controller.Config{
+		controller.AgentRateLimitMax: "-5",
+	},
+	expectError: `negative agent-ratelimit-max \(-5\) not valid`,
+}, {
+	about: "agent-ratelimit-rate missing unit",
+	config: controller.Config{
+		controller.AgentRateLimitRate: "150",
+	},
+	expectError: `agent-ratelimit-rate: conversion to duration: time: missing unit in duration "?150"?`,
+}, {
+	about: "agent-ratelimit-rate bad type, int",
+	config: controller.Config{
+		controller.AgentRateLimitRate: 150,
+	},
+	expectError: `agent-ratelimit-rate: expected string or time.Duration, got int\(150\)`,
+}, {
+	about: "agent-ratelimit-rate zero",
+	config: controller.Config{
+		controller.AgentRateLimitRate: "0s",
+	},
+	expectError: `agent-ratelimit-rate cannot be zero`,
+}, {
+	about: "agent-ratelimit-rate negative",
+	config: controller.Config{
+		controller.AgentRateLimitRate: "-5s",
+	},
+	expectError: `agent-ratelimit-rate cannot be negative`,
+}, {
+	about: "agent-ratelimit-rate too large",
+	config: controller.Config{
+		controller.AgentRateLimitRate: "4h",
+	},
+	expectError: `agent-ratelimit-rate must be between 0..1m`,
+}, {
+	about: "max-charm-state-size non-int",
+	config: controller.Config{
+		controller.MaxCharmStateSize: "ten",
+	},
+	expectError: `max-charm-state-size: expected number, got string\("ten"\)`,
+}, {
+	about: "max-charm-state-size cannot be negative",
+	config: controller.Config{
+		controller.MaxCharmStateSize: "-42",
+	},
+	expectError: `invalid max charm state size: should be a number of bytes \(or 0 to disable limit\), got -42`,
+}, {
+	about: "max-agent-state-size non-int",
+	config: controller.Config{
+		controller.MaxAgentStateSize: "ten",
+	},
+	expectError: `max-agent-state-size: expected number, got string\("ten"\)`,
+}, {
+	about: "max-agent-state-size cannot be negative",
+	config: controller.Config{
+		controller.MaxAgentStateSize: "-42",
+	},
+	expectError: `invalid max agent state size: should be a number of bytes \(or 0 to disable limit\), got -42`,
+}, {
+	about: "combined charm/agent state cannot exceed mongo's 16M limit/doc",
+	config: controller.Config{
+		controller.MaxCharmStateSize: "14000000",
+		controller.MaxAgentStateSize: "3000000",
+	},
+	expectError: `invalid max charm/agent state sizes: combined value should not exceed mongo's 16M per-document limit, got 17000000`,
+}, {
+	about: "public-dns-address: expect string, got number",
+	config: controller.Config{
+		controller.PublicDNSAddress: 42,
+	},
+	expectError: `public-dns-address: expected string, got int\(42\)`,
+}, {
+	about: "migration-agent-wait-time not a duration",
+	config: controller.Config{
+		controller.MigrationMinionWaitMax: "15",
+	},
+	expectError: `migration-agent-wait-time: conversion to duration: time: missing unit in duration "15"`,
+}, {
+	about: "application-resource-download-limit cannot be negative",
+	config: controller.Config{
+		controller.ApplicationResourceDownloadLimit: "-42",
+	},
+	expectError: `negative application-resource-download-limit \(-42\) not valid, use 0 to disable the limit`,
+}, {
+	about: "controller-resource-download-limit cannot be negative",
+	config: controller.Config{
+		controller.ControllerResourceDownloadLimit: "-42",
+	},
+	expectError: `negative controller-resource-download-limit \(-42\) not valid, use 0 to disable the limit`,
+}, {
+	about: "login token refresh url",
+	config: controller.Config{
+		controller.LoginTokenRefreshURL: `https://xxxx`,
+	},
+}, {
+	about: "invalid login token refresh url",
+	config: controller.Config{
+		controller.LoginTokenRefreshURL: `xxxx`,
+	},
+	expectError: `logic token refresh URL "xxxx" not valid`,
+}, {
+	about: "invalid query tracing value",
+	config: controller.Config{
+		controller.QueryTracingEnabled: "invalid",
+	},
+	expectError: `query-tracing-enabled: expected bool, got string\("invalid"\)`,
+}, {
+	about: "invalid query tracing threshold value",
+	config: controller.Config{
+		controller.QueryTracingThreshold: "invalid",
+	},
+	expectError: `query-tracing-threshold: conversion to duration: time: invalid duration "invalid"`,
+}, {
+	about: "negative query tracing threshold duration",
+	config: controller.Config{
+		controller.QueryTracingThreshold: "-1s",
+	},
+	expectError: `query-tracing-threshold value "-1s" must be a positive duration`,
+}}
 
 func (s *ConfigSuite) TestNewConfig(c *gc.C) {
 	for i, test := range newConfigTests {
@@ -820,6 +837,8 @@ func (s *ConfigSuite) TestDefaults(c *gc.C) {
 	c.Assert(cfg.ModelLogfileMaxSizeMB(), gc.Equals, controller.DefaultModelLogfileMaxSize)
 	c.Assert(cfg.ApplicationResourceDownloadLimit(), gc.Equals, controller.DefaultApplicationResourceDownloadLimit)
 	c.Assert(cfg.ControllerResourceDownloadLimit(), gc.Equals, controller.DefaultControllerResourceDownloadLimit)
+	c.Assert(cfg.QueryTracingEnabled(), gc.Equals, controller.DefaultQueryTracingEnabled)
+	c.Assert(cfg.QueryTracingThreshold(), gc.Equals, controller.DefaultQueryTracingThreshold)
 }
 
 func (s *ConfigSuite) TestAgentLogfile(c *gc.C) {
@@ -928,4 +947,28 @@ func (s *ConfigSuite) TestMigrationMinionWaitMax(c *gc.C) {
 
 	cfg[controller.MigrationMinionWaitMax] = "500ms"
 	c.Assert(cfg.MigrationMinionWaitMax(), gc.Equals, 500*time.Millisecond)
+}
+
+func (s *ConfigSuite) TestQueryTraceEnabled(c *gc.C) {
+	cfg, err := controller.NewConfig(
+		testing.ControllerTag.Id(),
+		testing.CACert, nil)
+	c.Assert(err, jc.ErrorIsNil)
+
+	c.Assert(cfg.QueryTracingEnabled(), gc.Equals, controller.DefaultQueryTracingEnabled)
+
+	cfg[controller.QueryTracingEnabled] = true
+	c.Assert(cfg.QueryTracingEnabled(), gc.Equals, true)
+}
+
+func (s *ConfigSuite) TestQueryTraceThreshold(c *gc.C) {
+	cfg, err := controller.NewConfig(
+		testing.ControllerTag.Id(),
+		testing.CACert, nil)
+	c.Assert(err, jc.ErrorIsNil)
+
+	c.Assert(cfg.QueryTracingThreshold(), gc.Equals, controller.DefaultQueryTracingThreshold)
+
+	cfg[controller.QueryTracingThreshold] = time.Second * 10
+	c.Assert(cfg.QueryTracingThreshold(), gc.Equals, time.Second*10)
 }
