+++ database/app/dqlite_other.go
@@ -62,6 +62,12 @@ func WithLogFunc(log client.LogFunc) Option {
 	return func() {}
 }
 
+// WithTracing will emit a log message at the given level every time a
+// statement gets executed.
+func WithTracing(level client.LogLevel) Option {
+	return func() {}
+}
+
 // App is a high-level helper for initializing a typical dqlite-based Go
 // application.
 //
